Fix DateUtils tests to use instance methods
